move ticketmaster class to components folder
