+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import SideBarMenu from './components/SideBarMenu';
 import Wrapper from './components/Wrapper';
-import Ticketmaster from './Ticketmaster';
+import Ticketmaster from './components/Ticketmaster';
 
 
 const App = () => (
